+++ app/Http/Controllers/Admin/PostController.php
@@ -4,16 +4,27 @@ namespace App\Http\Controllers\Admin;
 
 use App\pCategory;
 use App\post;
+use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Contracts\View\Factory;
 use Illuminate\Database\QueryException;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Routing\Redirector;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\View\View;
 
 class PostController extends Controller
 {
+    /**
+     * @return Factory|View
+     * @throws AuthorizationException
+     */
     public function create()
     {
+        $this->authorize('posts');
         $categories = pCategory::all();
         return view('admin.posts.create', [
             'categories' => $categories,
@@ -31,8 +42,14 @@ class PostController extends Controller
         ]);
     }
 
+    /**
+     * @param Request $request
+     * @return RedirectResponse|Redirector
+     * @throws AuthorizationException
+     */
     public function store(Request $request)
     {
+        $this->authorize('posts');
         $validate = $this->postValidation($request);
         if ($validate->fails())
             return back()->withErrors($validate, 'failed')->withInput();
@@ -53,8 +70,14 @@ class PostController extends Controller
     }
 
 
+    /**
+     * @param post $post
+     * @return Factory|View
+     * @throws AuthorizationException
+     */
     public function edit(post $post)
     {
+        $this->authorize('posts');
         $categories = pCategory::all();
         return view('admin.posts.edit', [
             'post' => $post,
@@ -62,8 +85,15 @@ class PostController extends Controller
         ]);
     }
 
+    /**
+     * @param post $post
+     * @param Request $request
+     * @return RedirectResponse|Redirector
+     * @throws AuthorizationException
+     */
     public function update(post $post, Request $request)
     {
+        $this->authorize('posts');
         $validate = $this->postValidation($request);
         if ($validate->fails())
             return back()->withErrors($validate, 'failed')->withInput();
@@ -75,16 +105,28 @@ class PostController extends Controller
         return redirect(route('admin.posts.index'))->withErrors(['عملیات با موفقیت انجام شد'], 'success');
     }
 
+    /**
+     * @return Factory|View
+     * @throws AuthorizationException
+     */
     public function index()
     {
+        $this->authorize('posts');
         $posts = post::latest()->get();
         return view('admin.posts.index', [
             'posts' => $posts
         ]);
     }
 
+    /**
+     * @param post $post
+     * @return RedirectResponse
+     * @throws AuthorizationException
+     * @throws Exception
+     */
     public function destroy(post $post)
     {
+        $this->authorize('posts');
         $post->delete();
         return back()->withErrors(['عملیات با موفقیت انجام شد'], 'success');
     }
