+++ app/Http/Controllers/Admin/ServiceController.php
@@ -4,22 +4,37 @@ namespace App\Http\Controllers\Admin;
 
 use App\Models\Product;
 use App\Service;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Routing\Redirector;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rule;
+use Illuminate\View\View;
 
 class ServiceController extends Controller
 {
+    /**
+     * @return Factory|View
+     * @throws AuthorizationException
+     */
     public function index()
     {
+        $this->authorize('services');
         $services = Service::all();
         return view('admin.services.index', ['services' => $services]);
     }
 
+    /**
+     * @return Factory|View
+     * @throws AuthorizationException
+     */
     public function create()
     {
+        $this->authorize('services');
         return view('admin.services.create');
     }
 
@@ -31,8 +46,14 @@ class ServiceController extends Controller
         ]);
     }
 
+    /**
+     * @param Request $request
+     * @return RedirectResponse|Redirector
+     * @throws AuthorizationException
+     */
     public function store(Request $request)
     {
+        $this->authorize('services');
         $validator = $this->storeRequest($request);
         if ($validator->fails())
             return back()->withErrors($validator, 'failed');
@@ -44,13 +65,26 @@ class ServiceController extends Controller
         return redirect(route('admin.service.index'))->withErrors(['عملیات با موفقیت انجام شد'], 'success');
     }
 
+    /**
+     * @param Service $service
+     * @return Factory|View
+     * @throws AuthorizationException
+     */
     public function edit(Service $service)
     {
+        $this->authorize('services');
         return view('admin.services.edit', ['service' => $service]);
     }
 
+    /**
+     * @param Service $service
+     * @param Request $request
+     * @return RedirectResponse|Redirector
+     * @throws AuthorizationException
+     */
     public function update(Service $service, Request $request)
     {
+        $this->authorize('services');
         $validator = $this->storeRequest($request);
         if ($validator->fails())
             return back()->withErrors($validator, 'failed');
@@ -61,14 +95,26 @@ class ServiceController extends Controller
         return redirect(route('admin.service.index'))->withErrors(['عملیات با موفقیت انجام شد'], 'success');
     }
 
+    /**
+     * @param Service $service
+     * @return RedirectResponse|Redirector
+     * @throws AuthorizationException
+     */
     public function destroy(Service $service)
     {
+        $this->authorize('services');
         $service->delete();
         return redirect(route('admin.service.index'))->withErrors(['عملیات با موفقیت انجام شد'], 'success');
     }
 
+    /**
+     * @param Service $service
+     * @return Factory|View
+     * @throws AuthorizationException
+     */
     public function products(Service $service)
     {
+        $this->authorize('serviceProducts');
         $products = Product::all();
         return view('admin.services.products', [
             'products' => $products,
@@ -76,8 +122,15 @@ class ServiceController extends Controller
         ]);
     }
 
+    /**
+     * @param Service $service
+     * @param Request $request
+     * @return RedirectResponse
+     * @throws AuthorizationException
+     */
     public function updateProducts(Service $service, Request $request)
     {
+        $this->authorize('serviceProducts');
         DB::table('product_services')->where('service_id', $service->id)->delete();
         foreach ($request->products as $product) {
             $productPaper = explode('-', $product);
